Koordinaten der einzelnen Runs nicht in die Merge-View mit übernehmen
