+++ src/app/components/display-merged-run/display-merged-run.component.html
@@ -3,5 +3,6 @@
         *ngIf="svgElements$ | async as svgElements"
         [svgElements]="svgElements.list"
         [canvasHeight]="svgElements.height"
+        [persistCoordinates]="false"
     ></app-canvas>
 </div>
